requirements.txt
appdirs==1.4.4
atomicwrites==1.4.0
blinker==1.9.0
CacheControl==0.14.2
cachelib==0.13.0
chardet==3.0.4
firebase-admin==6.6.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Session==0.8.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.0
google-api-python-client==2.157.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.2
google-cloud-firestore==2.20.0
google-cloud-storage==3.0.0
google-crc32c==1.6.0
google-generativeai==0.8.4
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
googletrans==4.0.0rc1
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
jsonpointer==2.1
langdetect==1.0.9
lxml==5.2.1
matplotlib==3.9.2
mkl-service==2.4.0
msgspec==0.19.0
proto-plus==1.25.0
protobuf==5.29.2
pyasn1-modules==0.2.8
pyls-spyder==0.4.0
PyQt5==5.15.10
PyQtWebEngine==5.15.6
pywin32==305.1
rfc3986==1.5.0
rsa==4.9
setuptools==75.1.0
uritemplate==4.1.1
Werkzeug==3.1.3
wheel==0.44.0